+++ src/pbox.h
@@ -4,5 +4,5 @@
 #define P_TABLE_SIZE 32
 #define E_TABLE_SIZE 48
 
-uint64_t expand(uint64_t block);
-uint64_t permute(uint64_t block);
+uint64_t expand(uint32_t block);
+uint32_t permute(uint32_t block);
